DEMO/EPASS/Image/requirements.txt
streamlit
numpy
Pillow
torch>=1.8.0 --index-url https://download.pytorch.org/whl/cpu
torchvision>=0.9.0 --index-url https://download.pytorch.org/whl/cpu